==> champs/3-1_merge_feature.ipynb <==
import pandas as pd
import numpy as np
import time
import random
import lightgbm as lgb
import time

import pickle
import matplotlib.pyplot as plt

from process import map_pred

FOLDER = '../../data_kaggle/champs/'
OUTPUT = FOLDER + 'out/'
TEMP = OUTPUT + 'temp/'
MAP  = OUTPUT + 'map/'

# original input
df_train = pd.read_csv(FOLDER+'train.csv')
df_test = pd.read_csv(FOLDER+'test.csv')
mols_train = df_train['molecule_name'].unique()

DATA_DATE = '20190721'
MAP_DATE = '20190728'

# input files
map_train = MAP + MAP_DATE + '_' + 'map_train.pickle'
map_test = MAP + MAP_DATE + '_' + 'map_test.pickle'

first_train = OUTPUT + DATA_DATE + '_' + 'dist_ang_ori_bond_cos_train_{}.pickle'
first_train_all = OUTPUT + DATA_DATE + '_' + 'features_train.pickle'

first_test = OUTPUT + DATA_DATE + '_' + 'dist_ang_ori_bond_cos_test_{}.pickle'
first_test_all = OUTPUT + DATA_DATE + '_' + 'features_test.pickle'

df_map_train = pd.read_pickle(map_train)
df_map_train.head()

df_features_train = pd.DataFrame()
bond_types = ['1JHC', '1JHN', '2JHH', '2JHC', '2JHN', '3JHH', '3JHC', '3JHN']
for b in bond_types:
    df_features_train = pd.concat([df_features_train, pd.read_pickle(first_train.format(b))], axis=0)
df_features_train = df_features_train.sort_values('id').reset_index(drop=True)
df_features_train.head()

df_features_train = df_features_train.fillna(0)
features_train = df_features_train[df_features_train.columns[6:]].values

with open(first_train_all, 'wb') as f:
    pickle.dump(features_train, f)

df_features_test = pd.DataFrame()
bond_types = ['1JHC', '1JHN', '2JHH', '2JHC', '2JHN', '3JHH', '3JHC', '3JHN']
for b in bond_types:
    df_features_test = pd.concat([df_features_test, pd.read_pickle(first_test.format(b))], axis=0)
df_features_test = df_features_test.sort_values('id').reset_index(drop=True)
df_features_test.head()

df_features_test = df_features_test.fillna(0)
features_test = df_features_test[df_features_test.columns[5:]].values

with open(first_test_all, 'wb') as f:
    pickle.dump(features_test, f)
